+++ src/app/core/learning-object-module/bundling/bundling.service.ts
@@ -3,6 +3,8 @@ import { Injectable } from '@angular/core';
 import { BUNDLING_ROUTES } from './bundling.routes';
 import { catchError } from 'rxjs/operators';
 import { throwError } from 'rxjs';
+import { LibraryService } from 'app/core/library-module/library.service';
+import { LearningObject } from 'entity/learning-object/learning-object';
 
 @Injectable({
   providedIn: 'root'
@@ -13,7 +15,12 @@ export class BundlingService {
 
   constructor(
     private http: HttpClient,
+    private libraryService: LibraryService
   ) { }
+
+  showDownloadModal = false;
+  currentIndex = null;
+  downloading = false;
 
   async bundleLearningObject(learningObjectId: string): Promise<void> {
     this.http
@@ -43,5 +50,49 @@ export class BundlingService {
       return throwError(error);
     }
   }
+
+  toggleBundle(
+    learningObjectId: string,
+    fileIDs: string[],
+    state: boolean
+  ) {
+    const route = BUNDLING_ROUTES.TOGGLE_BUNDLE_FILE({ learningObjectId });
+
+    return this.http
+      .patch(
+        route,
+        {
+          fileIDs: fileIDs,
+          packagable: state
+        },
+        { headers: this.headers, withCredentials: true }
+      )
+      .pipe(
+
+        catchError(this.handleError)
+      )
+      .toPromise();
+  }
+
+
+  toggleDownloadModal(val?: boolean) {
+    this.showDownloadModal = val;
+  }
+
+  download(learningObjectId: string) {
+    this.toggleDownloadModal(true);
+    this.libraryService.downloadBundle(BUNDLING_ROUTES.DOWNLOAD_BUNDLE(learningObjectId));
+  }
+
+
+  downloadObject(event: MouseEvent, object: LearningObject, index: number) {
+    event.stopPropagation();
+    this.currentIndex = index;
+    this.downloading[index] = true;
+    this.showDownloadModal = true;
+    this.libraryService.downloadBundle(BUNDLING_ROUTES.DOWNLOAD_BUNDLE(object.id));
+    this.downloading[index] = false;
+  }
+
 
 }
