+++ src/app/cube/library/library.component.ts
@@ -12,6 +12,7 @@ import { LearningObjectService } from '../learning-object.service';
 import { trigger, style, group, transition, animate, query } from '@angular/animations';
 import { NavbarService } from 'app/core/client-module/navbar.service';
 import { BUNDLING_ROUTES } from 'app/core/learning-object-module/bundling/bundling.routes';
+import { BundlingService } from 'app/core/learning-object-module/bundling/bundling.service';
 @Component({
   selector: 'clark-library',
   templateUrl: './library.component.html',
@@ -234,14 +235,14 @@ export class LibraryComponent implements OnInit, OnDestroy {
     }
   }
 
-  downloadObject(event: MouseEvent, object: LearningObject, index: number) {
-    event.stopPropagation();
-    this.currentIndex = index;
-    this.downloading[index] = true;
-    this.showDownloadModal = true;
-    this.libraryService.downloadBundle(BUNDLING_ROUTES.DOWNLOAD_BUNDLE(object.id));
-    this.downloading[index] = false;
-  }
+  // downloadObject(event: MouseEvent, object: LearningObject, index: number) {
+  //   event.stopPropagation();
+  //   this.currentIndex = index;
+  //   this.downloading[index] = true;
+  //   this.showDownloadModal = true;
+  //   this.libraryService.downloadBundle(BUNDLING_ROUTES.DOWNLOAD_BUNDLE(object.id));
+  //   this.downloading[index] = false;
+  // }
 
   // TODO: Come back and cry about this
   goToNotification(notification: any) {
