+++ src/app/onion/learning-object-builder/builder-store.service.ts
@@ -22,7 +22,7 @@ import { v4 as uuidv4 } from 'uuid';
 import { DirectoryNode } from 'app/shared/modules/filesystem/DirectoryNode';
 import { SubmissionsService } from 'app/core/learning-object-module/submissions/submissions.service';
 import { OutcomeService } from 'app/core/learning-object-module/outcomes/outcome.service';
-
+import { BundlingService } from 'app/core/learning-object-module/bundling/bundling.service';
 /**
  * Defines a list of actions the builder can take
  *
@@ -155,6 +155,7 @@ export class BuilderStore {
     private uriRetriever: UriRetrieverService,
     private submissionService: SubmissionsService,
     private outcomeService: OutcomeService,
+    private bundlingService: BundlingService
   ) {
     // subscribe to our objectCache$ observable and initiate calls to save object after a debounce
     this.objectCache$
@@ -250,18 +251,19 @@ export class BuilderStore {
     let revisionId: any = 0;
     // conditionally call either the getLearningObject function or the getLearningObjectRevision function based on function input
     const retrieve = this._isRevision && revisionId !== undefined && username ?
-    async () => {
-      // eslint-disable-next-line eqeqeq
-      if (revisionId == 0) {
-        revisionId = await this.learningObjectService.createRevision(cuid);
-      }
+      async () => {
+        // eslint-disable-next-line eqeqeq
+        if (revisionId == 0) {
+          revisionId = await this.learningObjectService.createRevision(cuid);
+        }
 
-      return this.learningObjectService.getLearningObjectRevision(username, cuid, revisionId);
-    } :
-    async () => {
-      const value = this.uriRetriever.getLearningObject({cuidInfo: {cuid, version}}, ['children', 'parents', 'materials', 'outcomes']);
-      return value.toPromise();
-    };
+        return this.learningObjectService.getLearningObjectRevision(username, cuid, revisionId);
+      } :
+      async () => {
+        const value = this.uriRetriever.getLearningObject({ cuidInfo: { cuid, version } }, ['children', 'parents',
+          'materials', 'outcomes']);
+        return value.toPromise();
+      };
 
     return retrieve()
       .then(object => {
@@ -461,14 +463,14 @@ export class BuilderStore {
   }) {
     if (event.item instanceof DirectoryNode) { // event.item is a Folder
       const fileIDs = this.getAllFolderFileIDs(event.item);
-      await this.learningObjectService.toggleBundle(
+      await this.bundlingService.toggleBundle(
         this.learningObject.id,
         fileIDs,
         event.state
       );
     } else { // event.item is a File
       const fileID = [event.item._id];
-      await this.learningObjectService.toggleBundle(
+      await this.bundlingService.toggleBundle(
         this.learningObject.id,
         fileID,
         event.state
@@ -533,11 +535,11 @@ export class BuilderStore {
     if (!checkIfUUID(outcome.serviceId || outcome.id) && (outcome.serviceId || outcome.id)) {
       this.serviceInteraction$.next(true);
       this.learningObjectService
-      .deleteOutcome(outcome.serviceId || outcome.id)
-      .then(() => {
-        this.serviceInteraction$.next(false);
-      })
-      .catch(e => this.handleServiceError(e, BUILDER_ERRORS.DELETE_OUTCOME));
+        .deleteOutcome(outcome.serviceId || outcome.id)
+        .then(() => {
+          this.serviceInteraction$.next(false);
+        })
+        .catch(e => this.handleServiceError(e, BUILDER_ERRORS.DELETE_OUTCOME));
     }
   }
 
@@ -822,8 +824,10 @@ export class BuilderStore {
     // Update the learning object
     this.learningObjectEvent.next(this.learningObject);
     this.saveObject({
-      materials: { folderDescriptions: this.learningObject.materials
-        .folderDescriptions }
+      materials: {
+        folderDescriptions: this.learningObject.materials
+          .folderDescriptions
+      }
     });
   }
 
@@ -901,7 +905,7 @@ export class BuilderStore {
     // Get most up-to-date values for current learning object
     await this.fetch(this.learningObject.cuid, this.learningObject.version);
     // Retrieve outcomes of current learning object
-    const value = await this.uriRetriever.getLearningObject({ cuidInfo: {cuid: this.learningObject.cuid} }, ['outcomes']).toPromise();
+    const value = await this.uriRetriever.getLearningObject({ cuidInfo: { cuid: this.learningObject.cuid } }, ['outcomes']).toPromise();
     // Iterate through outcomes
     value.outcomes.map(outcome => {
       // If the outcome text is empty, remove outcome
